+++ src/main/java/net/thesquire/backroomsmod/BackroomsModClient.java
@@ -1,7 +1,10 @@
 package net.thesquire.backroomsmod;
 
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.rendering.v1.BlockEntityRendererRegistry;
+import net.minecraft.client.render.RenderLayer;
+import net.thesquire.backroomsmod.block.ModBlocks;
 import net.thesquire.backroomsmod.block.entity.ModBlockEntities;
 import reborncore.client.multiblock.MultiblockRenderer;
 
@@ -10,6 +13,8 @@ public class BackroomsModClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         BlockEntityRendererRegistry.register(ModBlockEntities.MAGNETIC_DISTORTION_SYSTEM_CONTROL_COMPUTER, MultiblockRenderer::new);
+
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.MOUNTABLE_FLUORESCENT_LIGHT, RenderLayer.getCutout());
     }
 
 }